Fix rest controller

+++ src/main/java/com/api/student/repository/StudentRepository.java
@@ -4,6 +4,7 @@
 package com.api.student.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.stereotype.Repository;
 
 import com.api.student.entity.Student;
@@ -13,6 +14,6 @@ import com.api.student.entity.Student;
  *
  */
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer>{
+public interface StudentRepository extends JpaRepository<Student, Long>{
 
 }
